+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## unreleased
 * Add `billing_address` attribute reader to `ApplePayCard`
 * Add `pre_dispute_program` to `Dispute` and `DisputeSearch`
+* Add `AutoAccepted` status to `Dispute` 
+* Add `DisputeAutoAccepted` webhook to `WebhookNotification`
 
 ## 4.9.0
 * Add `cardholder_name` attribute reader to `ApplePayCard` (Thanks @AxelTheGerman)